No longer use Daemon.Runner to handle our daemonization.
We always run in foreground mode, and rely
on the end-user to kick us off in an Upstart
job or something to run as a daemon. We also
switch to a fully threaded model where each
'service' we register has its own Watcher
thread, and each Watcher thread spawns unique
Command threads to run a given service check.